requirements.txt
flask
flask-login
flask-sqlalchemy
flask_bcrypt
flask_wtf
wtforms
email_validator